move sideNav to its own component
